chore: move into utils
