+++ packages/policies/src/index.ts
@@ -1,2 +1,3 @@
 export * from "./policy";
 export * from "./types";
+export * from "./utils";
