+++ packages/policies/src/policy.ts
@@ -93,40 +93,3 @@ export class Policy<TResources extends Resources, TResourceIdentifier extends st
     return false;
   }
 }
-
-export const mergePolicies = <TResources extends Resources, TResourceIdentifier extends string = string>(
-  policyOne: Policy<TResources, TResourceIdentifier>,
-  policyTwo: Policy<TResources, TResourceIdentifier>,
-): Policy<TResources, TResourceIdentifier> => {
-
-  const map: Record<string, any> = {};
-
-  [policyOne, policyTwo].forEach((policy) => {
-    policy.statements.forEach(statement => {
-      for (const resourceType in statement.resources) {
-        if (!map[resourceType]) {
-          map[resourceType] = {};
-        }
-        for (const resourceId in statement.resources[resourceType]!) { // Non-null assertion added here
-          const resourceActions = statement.resources[resourceType]![resourceId]; // Non-null assertions added here
-          if (!map[resourceType][resourceId]) {
-            map[resourceType][resourceId] = new Set(resourceActions);
-          } else {
-            resourceActions.forEach((action: any) => {
-              (map[resourceType][resourceId] as Set<any>).add(action);
-            });
-          }
-        }
-      }
-    });
-  });
-
-  // Transform Set back into an array
-  for (const resourceType in map) {
-    for (const resourceId in map[resourceType]) {
-      map[resourceType][resourceId] = Array.from(map[resourceType][resourceId] as Set<any>);
-    }
-  }
-
-  return new Policy<TResources, TResourceIdentifier>({ resources: map as any });
-};
